Fix build errors and run bugs at patch 600/1433 of v4.11->v4.12 update
- Add trace functions
- Fix drm structs after upstream refactoring
- Cast arguments for clflush() calls
- Disable two new char* sysctls

+++ drivers/gpu/drm/i915/intel_ringbuffer.h
@@ -468,9 +468,9 @@ intel_write_status_page(struct intel_engine_cs *engine,
 			int reg, u32 value)
 {
 	mb();
-	clflush(&engine->status_page.page_addr[reg]);
+	clflush((u_long)&engine->status_page.page_addr[reg]);
 	engine->status_page.page_addr[reg] = value;
-	clflush(&engine->status_page.page_addr[reg]);
+	clflush((u_long)&engine->status_page.page_addr[reg]);
 	mb();
 }
